requirements.txt
absl-py==0.12.0
aiohttp==3.7.4.post0
alabaster==0.7.12
antlr4-python3-runtime==4.8
appdirs==1.4.4
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attrdict==2.0.1
attrs==21.2.0
audioread==2.1.9
Babel==2.9.1
backcall==0.2.0
black==19.10b0
bleach==3.3.0
boto3==1.17.83
botocore==1.20.83
braceexpand==0.1.7
cached-property==1.5.2
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
configparser==5.0.2
cycler==0.10.0
Cython==0.29.23
decorator==5.0.9
defusedxml==0.7.1
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.17.1
editdistance==0.5.3
entrypoints==0.3
filelock==3.0.12
flatbuffers==1.12
frozendict==2.0.2
fsspec==2021.5.0
future==0.18.2
g2p-en==2.1.0
gast==0.4.0
gdown==3.13.0
gitdb==4.0.7
GitPython==3.1.17
google-auth==1.30.1
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
huggingface-hub==0.0.8
hydra-core==1.0.6
idna==2.10
imagesize==1.2.0
importlib-metadata==4.3.0
importlib-resources==5.1.4
inflect==5.3.0
iniconfig==1.1.1
ipadic==1.0.0
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==4.3.21
jedi==0.18.0
jieba==0.42.1
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaldi-io==0.9.4
kaldi-python-io==1.2.2
kaldiio==2.17.2
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
latexcodec==2.0.1
librosa==0.8.1
llvmlite==0.36.0
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.12.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mecab-python3==1.0.3
megatron-lm==1.1.5
mistune==0.8.4
mpmath==1.2.1
multidict==5.1.0
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nemo-toolkit==1.0.0rc1
nest-asyncio==1.5.1
nltk==3.6.2
notebook==6.4.0
numba==0.53.1
numpy==1.19.5
oauthlib==3.1.0
omegaconf==2.0.6
onnx==1.9.0
OpenCC==1.1.2
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
pangu==4.0.6.1
parameterized==0.8.1
parso==0.8.2
pathspec==0.8.1
pathtools==0.1.2
pesq==0.0.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pip-api==0.0.20
pipreqs==0.4.10
pluggy==0.13.1
pooch==1.3.0
portalocker==2.0.0
prometheus-client==0.10.1
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.1
psutil==5.8.0
ptyprocess==0.7.0
py==1.10.0
pyannote.core==4.1
pyannote.database==4.1
pyannote.metrics==3.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.2
pybtex==0.24.0
pybtex-docutils==1.0.0
pycparser==2.20
pydub==0.25.1
Pygments==2.9.0
pyparsing==2.4.7
pypinyin==0.41.0
pyrsistent==0.17.3
PySocks==1.7.1
pystoi==0.3.3
pytest==6.2.4
pytest-runner==5.3.1
python-dateutil==2.8.1
pytorch-lightning==1.1.5
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
rapidfuzz==1.4.1
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
sacrebleu==1.5.1
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.6.3
Send2Trash==1.5.0
sentencepiece==0.1.95
sentry-sdk==1.1.0
shellingham==1.4.0
shortuuid==1.0.1
simplejson==3.17.2
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
sortedcontainers==2.4.0
SoundFile==0.10.3.post1
sox==1.4.1
spectralcluster==0.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-bibtex==2.2.1
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
subprocess32==3.5.4
sympy==1.8
tabulate==0.8.9
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
termcolor==1.1.0
terminado==0.10.0
testpath==0.5.0
threadpoolctl==2.1.0
tokenizers==0.10.3
toml==0.10.2
torch==1.7.1
torch-stft==0.1.4
torchaudio==0.7.2
torchtext==0.8.0
torchvision==0.8.2
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
transformers==4.6.1
typed-ast==1.4.3
typer==0.3.2
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.5
wandb==0.10.31
wcwidth==0.2.5
webdataset==0.1.62
webencodings==0.5.1
Werkzeug==2.0.1
wget==3.2
widgetsnbextension==3.5.1
wrapt==1.12.1
yarg==0.1.9
yarl==1.6.3
youtokentome==1.0.6
zipp==3.4.1
